oop-mock-common/src/main/java/com/codeheadsystems/oop/mock/factory/ObjectMapperFactory.java
// Copyright (c) 2022. CodeHead Systems. All rights reserved
// Ned Wolpert <ned.wolpert@codeheadsystems.com>

package com.codeheadsystems.oop.mock.factory;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
import javax.inject.Inject;
import javax.inject.Singleton;

@Singleton
public class ObjectMapperFactory {

    @Inject
    public ObjectMapperFactory() {

    }

    public ObjectMapper objectMapper() {
        return new ObjectMapper()
                .registerModule(new Jdk8Module());
    }

}
